feat(home): make the lobby the homepage
The lobby has been living at /lobby as an alternate homepage since it shipped
(R38), with promotion to / called out as a separate, explicitly-approved flip.
This is that flip.

/ is a client-side switcher that also serves /terms-of-service,
/privacy-policy and /support-the-pod, so only the 'landing' view changes;
the legal views are untouched. /lobby keeps working — both routes render the
same component — so existing links and bookmarks don't break.

Swapping the pages naively would have quietly dropped things the old homepage
owned and the lobby never had:

  - the footer: Stats/QA/API/GitHub/Patreon/Swag/Companion and, more to the
    point, Support the Pod, Terms and Privacy — taking the legal links off the
    site's front door.
  - the FFG/Disney disclaimer.
  - <ReleaseNotes />, which is how anyone finds out what shipped.

So the footer is extracted to a shared SiteFooter used by BOTH pages (one copy,
so the links can't drift), and the lobby now renders ReleaseNotes too.

SiteFooter sits outside .lobby-page, which is capped at 100vh: the lobby keeps
its no-fold guarantee and the footer sits below it. In practice the vh-clamped
lobby leaves room for both — verified zero page overflow at 1440x900 (lobby
686 + footer 74) and 1280x700 (616 + 74), footer fully visible at both.
